+++ registry.go
@@ -1,19 +1,20 @@
 package distribution
 
 import (
+	"io"
+	"time"
+
 	"github.com/docker/distribution/digest"
 	"github.com/docker/distribution/manifest"
 	"golang.org/x/net/context"
 )
-
-// TODO(stevvooe): These types need to be moved out of the storage package.
 
 // Registry represents a collection of repositories, addressable by name.
 type Registry interface {
 	// Repository should return a reference to the named repository. The
 	// registry may or may not have the repository but should always return a
 	// reference.
-	Repository(ctx context.Context, name string) Repository
+	Repository(ctx context.Context, name string) (Repository, error)
 }
 
 // Repository is a named collection of manifests and layers.
@@ -81,4 +82,43 @@ type LayerService interface {
 	// upload. The caller should seek to the latest desired upload location
 	// before proceeding.
 	Resume(uuid string) (LayerUpload, error)
+}
+
+// Layer provides a readable and seekable layer object. Typically,
+// implementations are *not* goroutine safe.
+type Layer interface {
+	// http.ServeContent requires an efficient implementation of
+	// ReadSeeker.Seek(0, os.SEEK_END).
+	io.ReadSeeker
+	io.Closer
+
+	// Digest returns the unique digest of the blob, which is the tarsum for
+	// layers.
+	Digest() digest.Digest
+
+	// CreatedAt returns the time this layer was created.
+	CreatedAt() time.Time
+}
+
+// LayerUpload provides a handle for working with in-progress uploads.
+// Instances can be obtained from the LayerService.Upload and
+// LayerService.Resume.
+type LayerUpload interface {
+	io.WriteSeeker
+	io.ReaderFrom
+	io.Closer
+
+	// UUID returns the identifier for this upload.
+	UUID() string
+
+	// StartedAt returns the time this layer upload was started.
+	StartedAt() time.Time
+
+	// Finish marks the upload as completed, returning a valid handle to the
+	// uploaded layer. The digest is validated against the contents of the
+	// uploaded layer.
+	Finish(digest digest.Digest) (Layer, error)
+
+	// Cancel the layer upload process.
+	Cancel() error
 }
